// src/types/resources/Communication.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCanonical } from '../simpleTypes/Canonical';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TDateTime } from '../simpleTypes/DateTime';
import { TCommunicationPayload } from '../partials/CommunicationPayload';
import { TAnnotation } from '../partials/Annotation';

export type TCommunication = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  instantiatesCanonical?: TCanonical[];
  instantiatesUri?: TUri[];
  basedOn?: TReference[];
  partOf?: TReference[];
  inResponseTo?: TReference[];
  status: string;
  statusReason?: TCodeableConcept;
  category?: TCodeableConcept[];
  priority?: string;
  medium?: TCodeableConcept[];
  subject?: TReference;
  topic?: TCodeableConcept;
  about?: TReference[];
  encounter?: TReference;
  sent?: TDateTime;
  received?: TDateTime;
  recipient?: TReference[];
  sender?: TReference;
  reasonCode?: TCodeableConcept[];
  reasonReference?: TReference[];
  payload?: TCommunicationPayload[];
  note?: TAnnotation[];
};